[fd/hls] Support `--write-pages` for m3u8 media playlists
Authored by: bashonly
